net: shell: Add events command
Move "events" command to a separate file.
