+++ subsys/net/lib/shell/CMakeLists.txt
@@ -11,6 +11,7 @@ zephyr_library_sources(arp.c)
 zephyr_library_sources(capture.c)
 zephyr_library_sources(conn.c)
 zephyr_library_sources(dns.c)
+zephyr_library_sources(events.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
